Bugfix "when open the dialog of a subproject and SS is down, the dialog is not being displayed"
The code would try and get the possible data-release-strategy options from SS (even though
you can't edit the strategy in the subproject) and there was nothing dealing with the case
where it can't get that information.
Now if it can't get through to SS, it returns the information it already had
(the cached study name in the database, or "No strategy" as appropriate).